Kristal toplama düzeltmesi ve ana geminin yeniden çizimi
Kristaller toplanmıyordu — iki ayrı neden:

- Toplayıcı kargoyu tam birime yuvarlıyor, artan kesri ayrı bir birikeçte
  tutuyor ve hangarda boşaltırken sıfırlıyordu. Level 1'de bir asteroit
  parçası 0.5 kaynak düşürdüğü için kristalin tamamı yanıyordu. Kargo ve
  ResourceInventory.Add artık float.
- Kristal, metalle aynı miktarı paylaşıyordu; bu hiç gerekçelendirilmemişti,
  tek kod yolunu paylaşmalarından düşmüştü. 0.5 birim × %12 olasılık = tam
  parçalanmış büyük asteroit başına 0.37 kristal. Kristal artık kendi
  tabanını taşıyor (CrystalMinAmount = 3, asteroit başına ~2.3).

Enkaz görünürlüğü: kristal parlak camgöbeği ve daha iri çiziliyor. Eski
mavimsi gri, ekranda ~7 pikselken kahverengi metalden ayırt edilemiyordu.

Ana gemi 4x1 → 4.0x2.4 birim (tuval 1600x960) ve yan profil olarak yeniden
çizildi. Eski siluet iki ucu da sivrilen bir mercekti — bir gövdenin plan
görünümünün işareti, deniz gemisi gibi okunuyordu.

Slotlar artık ızgara değil, gövdenin yapılarını takip ediyor: kıç makine
bloğu (0/3/7, arkasında 4 itici), sırt kulesi (1 ana silah, 4), bel gövdesi
(5/8), karın hangar modülü (6, bindirme ağızları), baş kesimi (2/9). Izgarayı
koruyup gemiyi büyütmek denendi ve silueti tuğlaya çeviriyordu; slotları
taşımak pruva kaması ve kıç basamağını mümkün kıldı. StartingLoadout'un
slot numaraları korundu, yeni düzende zaten doğru yerlere denk geliyor.

HealthBar geometrisi artık sprite bounds'undan türüyor — sahnedeki sabit
barOffsetY yeni gövdenin içinde kalıyordu.

DİKKAT: gövde collider'ı ~4x0.63 → ~4x2.2. Kalkan açıkken fark yok (mermi
önce 2.5 birimlik kalkan küresine çarpıyor) ama kalkan kapalıyken kesit
alanı ~3x büyüdü — oynanarak ölçülecek.

+++ StarfarerPrototype/Assets/Scripts/Ship/ResourceInventory.cs
@@ -66,7 +66,12 @@ public class ResourceInventory : MonoBehaviour
         }
     }
 
-    public void Add(ResourceType type, int amount)
+    /// <summary>
+    /// Miktar KESİRLİ tutulur. Eskiden int'ti ve toplayıcı kargosunu tam birime
+    /// yuvarlamak zorundaydı: level 1'de asteroit parçası 0.5 kaynak düşürdüğü
+    /// için kristalin tamamı yuvarlamada kayboluyordu.
+    /// </summary>
+    public void Add(ResourceType type, float amount)
     {
         switch (type)
         {
